Correct DoH base64 alphabet to base64url not base64
DoH is supposed to require the base64url encoding of the DNS
request, but our encoder was actually using classic base64. While
this is incorrect, it appears many (most?) DoH resolvers will
actually accept base64 just fine.

Still, best to swap for the correct alphabet, which we do here.

Fixes #12